+++ apps/web/src/state/farms/index.ts
@@ -1,6 +1,6 @@
+import { ChainId } from '@pancakeswap/chains'
 import { createFarmFetcher, SerializedFarm, SerializedFarmsState } from '@pancakeswap/farms'
 import { getFarmConfig } from '@pancakeswap/farms/constants'
-import { ChainId } from '@pancakeswap/chains'
 import { createAsyncThunk, createSlice, isAnyOf } from '@reduxjs/toolkit'
 import type {
   UnknownAsyncThunkFulfilledAction,
@@ -12,8 +12,8 @@ import stringify from 'fast-json-stable-stringify'
 import keyBy from 'lodash/keyBy'
 import type { AppState } from 'state'
 import { verifyBscNetwork } from 'utils/verifyBscNetwork'
-import { chains } from 'utils/wagmi'
 import { getViemClients } from 'utils/viem'
+import { chains } from 'utils/wagmi'
 import splitProxyFarms from 'views/Farms/components/YieldBooster/helpers/splitProxyFarms'
 import { Address } from 'wagmi'
 import { resetUserState } from '../global/actions'
@@ -31,7 +31,7 @@ const fetchFarmPublicDataPkg = async ({
   chain,
 }): Promise<[SerializedFarm[], number, number, string]> => {
   const farmsConfig = await getFarmConfig(chainId)
-  const farmsCanFetch = farmsConfig.filter((farmConfig) => pids.includes(farmConfig.pid))
+  const farmsCanFetch = farmsConfig?.filter((farmConfig) => pids.includes(farmConfig.pid)) ?? []
   const priceHelperLpsConfig = getFarmsPriceHelperLpFiles(chainId)
 
   const { farmsWithPrice, poolLength, regularCakePerBlock, totalRegularAllocPoint } = await farmFetcher.fetchFarms({
@@ -46,7 +46,7 @@ export const farmFetcher = createFarmFetcher(getViemClients)
 
 const initialState: SerializedFarmsState = {
   data: [],
-  chainId: null,
+  chainId: undefined,
   loadArchivedFarmsData: false,
   userDataLoaded: false,
   totalRegularAllocPoint: '0',
@@ -63,15 +63,16 @@ export const fetchInitialFarmsData = createAsyncThunk<
 >('farms/fetchInitialFarmsData', async ({ chainId }) => {
   return getFarmConfig(chainId).then((farmDataList) => {
     return {
-      data: farmDataList.map((farm) => ({
-        ...farm,
-        userData: {
-          allowance: '0',
-          tokenBalance: '0',
-          stakedBalance: '0',
-          earnings: '0',
-        },
-      })),
+      data:
+        farmDataList?.map((farm) => ({
+          ...farm,
+          userData: {
+            allowance: '0',
+            tokenBalance: '0',
+            stakedBalance: '0',
+            earnings: '0',
+          },
+        })) ?? [],
       chainId,
     }
   })
@@ -199,9 +200,8 @@ export const fetchFarmUserDataAsync = createAsyncThunk<
     }
     const poolLength = state.farms.poolLength ?? (await fetchMasterChefFarmPoolLength(ChainId.BSC))
     const farmsConfig = await getFarmConfig(chainId)
-    const farmsCanFetch = farmsConfig.filter(
-      (farmConfig) => pids.includes(farmConfig.pid) && poolLength > farmConfig.pid,
-    )
+    const farmsCanFetch =
+      farmsConfig?.filter((farmConfig) => pids.includes(farmConfig.pid) && poolLength > farmConfig.pid) ?? []
     if (proxyAddress && farmsCanFetch?.length && verifyBscNetwork(chainId)) {
       const { normalFarms, farmsWithProxy } = splitProxyFarms(farmsCanFetch)