+++ apps/web/src/views/CakeStaking/hooks/useContractWrite/useWriteIncreaseLockWeeksCallback.tsx
@@ -43,10 +43,14 @@ export const useWriteIncreaseLockWeeksCallback = () => {
     setTxHash(hash ?? '')
     setStatus(ApproveAndLockStatus.INCREASE_WEEKS_PENDING)
     if (hash) {
-      await waitForTransaction({ hash })
-      setCakeLockWeeks('')
+      const transactionReceipt = await waitForTransaction({ hash })
+      if (transactionReceipt?.status === 'success') {
+        setCakeLockWeeks('')
+        setStatus(ApproveAndLockStatus.CONFIRMED)
+      } else {
+        setStatus(ApproveAndLockStatus.ERROR)
+      }
     }
-    setStatus(ApproveAndLockStatus.CONFIRMED)
   }, [
     cakeLockWeeks,
     veCakeContract.simulate,