+++ apps/web/src/views/Home/components/Banners/PerpetualApolloXCampaignBanner.tsx
@@ -154,7 +154,7 @@ const PerpetualBanner = () => {
   const perpetualUrl = useMemo(() => getPerpetualUrl({ chainId, languageCode: code, isDark }), [chainId, code, isDark])
   const [onUSCitizenModalPresent] = useModal(
     <USCitizenConfirmModal title={t('PancakeSwap Perpetuals')} id={IdType.PERPETUALS} />,
-    false,
+    true,
     false,
     'usCitizenConfirmModal',
   )